+++ libs/ui/src/components/Explorer/manage-view-settings/filter-items/filter-type/BooleanAttributeDropdown.tsx
@@ -1,17 +1,15 @@
 // Copyright LEAV Solutions 2017 until 2023/11/05, Copyright Aristid from 2023/11/06
 // This file is released under LGPL V3
 // License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-import {useSharedTranslation} from '../../../../../hooks/useSharedTranslation';
-import {KitButton, KitDivider, KitSpace, KitSwitch} from 'aristid-ds';
 import {ComponentProps, FunctionComponent} from 'react';
-import {FaClock, FaTrash} from 'react-icons/fa';
+import {KitDivider, KitSpace, KitSwitch} from 'aristid-ds';
+import {AttributeConditionFilter} from '_ui/types';
+import {IExplorerFilter, IFilterDropDownProps} from '_ui/components/Explorer/_types';
 import {useViewSettingsContext} from '../../store-view-settings/useViewSettingsContext';
 import {ViewSettingsActionTypes} from '../../store-view-settings/viewSettingsReducer';
-import {IExplorerFilter, IFilterDropDownProps} from '../../../_types';
-import {AttributeConditionFilter} from '../../../../../types';
+import {FilterOptionsInDropDown} from './filter-options/FilterOptionsInDropDown';
 
 export const BooleanAttributeDropDown: FunctionComponent<IFilterDropDownProps> = ({filter}) => {
-    const {t} = useSharedTranslation();
     const {dispatch} = useViewSettingsContext();
 
     const _updateFilter = (filterData: IExplorerFilter) => {
@@ -25,32 +23,11 @@ export const BooleanAttributeDropDown: FunctionComponent<IFilterDropDownProps> =
         _updateFilter({...filter, condition: AttributeConditionFilter.EQUAL, value: checked ? 'true' : 'false'});
     };
 
-    const _onDeleteFilter: ComponentProps<typeof KitButton>['onClick'] = () =>
-        dispatch({
-            type: ViewSettingsActionTypes.REMOVE_FILTER,
-            payload: {
-                id: filter.id
-            }
-        });
-
-    const _onResetFilter: ComponentProps<typeof KitButton>['onClick'] = () =>
-        dispatch({
-            type: ViewSettingsActionTypes.RESET_FILTER,
-            payload: {
-                id: filter.id
-            }
-        });
-
     return (
         <KitSpace size="xxs" direction="vertical">
             <KitSwitch defaultValue={Boolean(filter.value)} onChange={_onBooleanChanged} />
             <KitDivider noMargin />
-            <KitButton type="redirect" icon={<FaClock />} onClick={_onResetFilter}>
-                {t('explorer.reset-filter')}
-            </KitButton>
-            <KitButton type="redirect" icon={<FaTrash />} onClick={_onDeleteFilter}>
-                {t('global.delete')}
-            </KitButton>
+            <FilterOptionsInDropDown filter={filter} />
         </KitSpace>
     );
 };
