+++ libs/ui/src/components/Explorer/manage-view-settings/filter-items/filter-type/DateRangeAttributeDropDown.tsx
@@ -1,15 +1,14 @@
 // Copyright LEAV Solutions 2017 until 2023/11/05, Copyright Aristid from 2023/11/06
 // This file is released under LGPL V3
 // License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-import {useSharedTranslation} from '_ui/hooks/useSharedTranslation';
-import {KitButton, KitDivider, KitSelect, KitSpace} from 'aristid-ds';
 import {ComponentProps, FunctionComponent} from 'react';
-import {FaClock, FaTrash} from 'react-icons/fa';
-import {AttributeConditionFilter} from '_ui/types';
-import {IExplorerFilter, IFilterDropDownProps} from '../../../_types';
+import {KitDivider, KitSelect, KitSpace} from 'aristid-ds';
+import {useSharedTranslation} from '_ui/hooks/useSharedTranslation';
+import {IExplorerFilter, IFilterDropDownProps} from '_ui/components/Explorer/_types';
 import {useViewSettingsContext} from '../../store-view-settings/useViewSettingsContext';
 import {ViewSettingsActionTypes} from '../../store-view-settings/viewSettingsReducer';
 import {useConditionsOptionsByType} from './useConditionOptionsByType';
+import {FilterOptionsInDropDown} from './filter-options/FilterOptionsInDropDown';
 
 export const DateRangeAttributeDropDown: FunctionComponent<IFilterDropDownProps> = ({filter}) => {
     const {t} = useSharedTranslation();
@@ -27,22 +26,6 @@ export const DateRangeAttributeDropDown: FunctionComponent<IFilterDropDownProps>
     const _onConditionChanged: ComponentProps<typeof KitSelect>['onChange'] = condition =>
         _updateFilter({...filter, condition});
 
-    const _onResetFilter: ComponentProps<typeof KitButton>['onClick'] = () =>
-        dispatch({
-            type: ViewSettingsActionTypes.RESET_FILTER,
-            payload: {
-                id: filter.id
-            }
-        });
-
-    const _onDeleteFilter: ComponentProps<typeof KitButton>['onClick'] = () =>
-        dispatch({
-            type: ViewSettingsActionTypes.REMOVE_FILTER,
-            payload: {
-                id: filter.id
-            }
-        });
-
     return (
         <KitSpace size="xxs" direction="vertical">
             <KitSelect
@@ -52,12 +35,7 @@ export const DateRangeAttributeDropDown: FunctionComponent<IFilterDropDownProps>
                 placeholder={t('explorer.select-condition')}
             />
             <KitDivider noMargin />
-            <KitButton type="redirect" icon={<FaClock />} onClick={_onResetFilter}>
-                {t('explorer.reset-filter')}
-            </KitButton>
-            <KitButton type="redirect" icon={<FaTrash />} onClick={_onDeleteFilter}>
-                {t('global.delete')}
-            </KitButton>
+            <FilterOptionsInDropDown filter={filter} />
         </KitSpace>
     );
 };
