+++ libs/ui/src/components/Explorer/manage-view-settings/filter-items/filter-type/SimpleFilterDropDown.tsx
@@ -1,14 +1,13 @@
 // Copyright LEAV Solutions 2017 until 2023/11/05, Copyright Aristid from 2023/11/06
 // This file is released under LGPL V3
 // License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
-import {useSharedTranslation} from '_ui/hooks/useSharedTranslation';
-import {KitButton, KitDivider, KitSelect, KitSpace} from 'aristid-ds';
 import {ComponentProps, FunctionComponent} from 'react';
-import {FaClock, FaTrash} from 'react-icons/fa';
-import {IExplorerFilter, IFilterDropDownProps} from '../../../_types';
+import {KitDivider, KitSelect, KitSpace} from 'aristid-ds';
+import {IExplorerFilter, IFilterDropDownProps} from '_ui/components/Explorer/_types';
 import {useViewSettingsContext} from '../../store-view-settings/useViewSettingsContext';
 import {ViewSettingsActionTypes} from '../../store-view-settings/viewSettingsReducer';
 import {FilterValueList} from './FilterValueList';
+import {FilterOptionsInDropDown} from './filter-options/FilterOptionsInDropDown';
 
 // TODO : This is an exemple file showing ho to customize dropdown Panel content. Don't mind the content of the file, missing types,... it's just an example.
 
@@ -46,7 +45,6 @@ const conditions = [
 const attributeValuesList = ['toto', 'tata', 'Value 3', 'Value 4', 'Value 5'];
 
 export const SimpleFilterDropdown: FunctionComponent<IFilterDropDownProps> = ({filter}) => {
-    const {t} = useSharedTranslation();
     const {dispatch} = useViewSettingsContext();
 
     const _updateFilter = (filterData: IExplorerFilter) => {
@@ -64,22 +62,6 @@ export const SimpleFilterDropdown: FunctionComponent<IFilterDropDownProps> = ({f
         _updateFilter({...filter, value: value.join('-')});
     };
 
-    const _onResetFilter: ComponentProps<typeof KitButton>['onClick'] = () =>
-        dispatch({
-            type: ViewSettingsActionTypes.RESET_FILTER,
-            payload: {
-                id: filter.id
-            }
-        });
-
-    const _onDeleteFilter: ComponentProps<typeof KitButton>['onClick'] = () =>
-        dispatch({
-            type: ViewSettingsActionTypes.REMOVE_FILTER,
-            payload: {
-                id: filter.id
-            }
-        });
-
     return (
         <KitSpace size="xxs" direction="vertical">
             <KitSelect options={conditions} onChange={_onConditionChanged} value={filter.condition} />
@@ -91,12 +73,7 @@ export const SimpleFilterDropdown: FunctionComponent<IFilterDropDownProps> = ({f
                 onSelectionChanged={_onValueClick}
             />
             <KitDivider noMargin />
-            <KitButton type="redirect" icon={<FaClock />} onClick={_onResetFilter}>
-                {t('explorer.reset-filter')}
-            </KitButton>
-            <KitButton type="redirect" icon={<FaTrash />} onClick={_onDeleteFilter}>
-                {t('global.delete')}
-            </KitButton>
+            <FilterOptionsInDropDown filter={filter} />
         </KitSpace>
     );
 };
