+++ libs/ui/src/components/Explorer/manage-view-settings/filter-items/filter-type/filter-options/FilterOptionsInDropDown.tsx
@@ -1,0 +1,42 @@
+// Copyright LEAV Solutions 2017 until 2023/11/05, Copyright Aristid from 2023/11/06
+// This file is released under LGPL V3
+// License text available at https://www.gnu.org/licenses/lgpl-3.0.txt
+import {ComponentProps, FunctionComponent} from 'react';
+import {FaClock, FaTrash} from 'react-icons/fa';
+import {KitButton} from 'aristid-ds';
+import {useSharedTranslation} from '_ui/hooks/useSharedTranslation';
+import {IFilterDropDownProps} from '_ui/components/Explorer/_types';
+import {ViewSettingsActionTypes} from '../../../store-view-settings/viewSettingsReducer';
+import {useViewSettingsContext} from '../../../store-view-settings/useViewSettingsContext';
+
+export const FilterOptionsInDropDown: FunctionComponent<IFilterDropDownProps> = ({filter}) => {
+    const {t} = useSharedTranslation();
+    const {dispatch} = useViewSettingsContext();
+
+    const _onResetFilter: ComponentProps<typeof KitButton>['onClick'] = () =>
+        dispatch({
+            type: ViewSettingsActionTypes.RESET_FILTER,
+            payload: {
+                id: filter.id
+            }
+        });
+
+    const _onDeleteFilter: ComponentProps<typeof KitButton>['onClick'] = () =>
+        dispatch({
+            type: ViewSettingsActionTypes.REMOVE_FILTER,
+            payload: {
+                id: filter.id
+            }
+        });
+
+    return (
+        <>
+            <KitButton type="redirect" icon={<FaClock />} onClick={_onResetFilter}>
+                {t('explorer.reset-filter')}
+            </KitButton>
+            <KitButton type="redirect" icon={<FaTrash />} onClick={_onDeleteFilter}>
+                {t('global.delete')}
+            </KitButton>
+        </>
+    );
+};
